# docs/demo-materials/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "电商订单中心 API",
    "version": "1.0.0",
    "description": "e2e-demo-project 的对外接口文档，用于测试【接口文档解析】与【AI 评审】。覆盖用户、订单、商品 3 个模块共 12 个接口。",
    "contact": { "name": "电商中台团队", "email": "order-team@example.com" }
  },
  "servers": [
    { "url": "https://api.example.com/v1", "description": "生产环境" },
    { "url": "https://staging.example.com/v1", "description": "预发布环境" }
  ],
  "tags": [
    { "name": "User", "description": "用户与鉴权" },
    { "name": "Order", "description": "订单管理" },
    { "name": "Product", "description": "商品中心" }
  ],
  "paths": {
    "/api/v1/users/register": {
      "post": {
        "tags": ["User"],
        "summary": "用户注册",
        "description": "手机号 + 密码注册，注册成功返回用户 ID，需先通过短信验证码。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterReq" },
              "example": {
                "phone": "13800138000",
                "password": "Passw0rd!",
                "sms_code": "654321",
                "nickname": "测试用户"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResp" }
              }
            }
          },
          "400": { "description": "参数错误或验证码失效" },
          "409": { "description": "手机号已注册" }
        },
        "security": []
      }
    },
    "/api/v1/users/login": {
      "post": {
        "tags": ["User"],
        "summary": "账号密码登录",
        "description": "使用手机号/邮箱 + 密码登录，返回 access_token（JWT，2h 过期）和 refresh_token（30 天过期）。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginReq" },
              "example": { "account": "13800138000", "password": "Passw0rd!" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功，返回双 Token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResp" }
              }
            }
          },
          "401": { "description": "账号或密码错误（连续 5 次将锁定 30 分钟）" }
        },
        "security": []
      }
    },
    "/api/v1/users/refresh": {
      "post": {
        "tags": ["User"],
        "summary": "刷新 Token",
        "description": "用 refresh_token 换新的 access_token，需保证 refresh_token 未过期且未被列入黑名单。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "refresh_token": { "type": "string" } },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResp" }
              }
            }
          },
          "401": { "description": "refresh_token 过期或失效" }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": ["User"],
        "summary": "查询当前用户信息",
        "description": "需要 Bearer Token，返回当前登录用户的完整资料。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResp" }
              }
            }
          },
          "401": { "description": "未登录或 token 过期" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/v1/users/password": {
      "put": {
        "tags": ["User"],
        "summary": "修改密码",
        "description": "登录态下使用旧密码换新密码。新密码强度要求：≥8 位、含大小写字母 + 数字 + 特殊字符。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordReq" }
            }
          }
        },
        "responses": {
          "200": { "description": "修改成功，15 分钟后旧 token 全部失效" },
          "400": { "description": "新密码不符合强度要求" },
          "401": { "description": "未登录" },
          "403": { "description": "旧密码错误" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": ["Product"],
        "summary": "商品列表查询",
        "description": "支持关键词搜索、分类筛选、价格区间、分页与排序。仅展示已上架商品。",
        "parameters": [
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "商品名称 / SKU 模糊匹配" },
          { "name": "category_id", "in": "query", "schema": { "type": "integer" }, "description": "分类 ID" },
          { "name": "min_price", "in": "query", "schema": { "type": "number", "minimum": 0 }, "description": "最低价（元）" },
          { "name": "max_price", "in": "query", "schema": { "type": "number", "minimum": 0 }, "description": "最高价（元）" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["price_asc", "price_desc", "newest", "sales"] }, "description": "排序方式" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageRespProduct" }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["Product"],
        "summary": "商品详情",
        "description": "根据 ID 查询商品详情，包含 SPU 信息、SKU 列表、库存与评价数。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "商品 ID（UUID）" }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductDetailResp" } } } },
          "404": { "description": "商品不存在或已下架" }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "tags": ["Order"],
        "summary": "创建订单",
        "description": "根据购物车 SKU 创建待支付订单。需要登录态。同一用户 1 分钟内最多创建 10 个订单（限流）。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderReq" },
              "example": {
                "items": [
                  { "sku_id": "SKU-001", "quantity": 2 },
                  { "sku_id": "SKU-002", "quantity": 1 }
                ],
                "address_id": "ADDR-1001",
                "coupon_id": null,
                "remark": "请尽快发货"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "订单创建成功，返回订单号", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderResp" } } } },
          "400": { "description": "请求参数错误（如库存不足、SKU 下架）" },
          "401": { "description": "未登录" },
          "429": { "description": "触发限流，请稍后再试" }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "get": {
        "tags": ["Order"],
        "summary": "订单列表查询",
        "description": "查询当前用户的订单列表，支持按状态 / 时间筛选。",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["PENDING", "PAID", "SHIPPED", "COMPLETED", "CANCELLED"] } },
          { "name": "start_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "end_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageRespOrder" } } } },
          "401": { "description": "未登录" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": ["Order"],
        "summary": "订单详情",
        "description": "根据订单 ID 查询订单详情，仅订单所属用户可访问。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderResp" } } } },
          "403": { "description": "无权访问该订单" },
          "404": { "description": "订单不存在" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/v1/orders/{id}/pay": {
      "post": {
        "tags": ["Order"],
        "summary": "订单支付",
        "description": "对指定订单发起支付。订单必须是当前用户的、状态为 PENDING。支付成功后状态变为 PAID，超时未支付系统自动取消（30 分钟）。该接口为幂等接口，重复调用不会重复扣款。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "idempotency_key", "in": "header", "required": true, "schema": { "type": "string" }, "description": "幂等键（UUID），相同 key 5 分钟内只会扣一次款" }
        ],
        "responses": {
          "200": { "description": "支付成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderResp" } } } },
          "400": { "description": "订单状态不允许支付" },
          "401": { "description": "未登录" },
          "403": { "description": "非订单所有者" },
          "404": { "description": "订单不存在" },
          "409": { "description": "幂等键冲突，请等待 5 分钟" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "tags": ["Order"],
        "summary": "取消订单",
        "description": "取消 PENDING 或 PAID 状态的订单。PAID 订单取消后会触发原路退款（1-3 个工作日）。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "reason", "in": "query", "schema": { "type": "string", "enum": ["USER_CANCEL", "OUT_OF_STOCK", "OTHER"] }, "description": "取消原因" }
        ],
        "responses": {
          "200": { "description": "取消成功，PAID 状态将进入退款流程" },
          "400": { "description": "当前状态不允许取消（如已完成 / 已退款）" },
          "401": { "description": "未登录" },
          "403": { "description": "非订单所有者" },
          "404": { "description": "订单不存在" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterReq": {
        "type": "object",
        "required": ["phone", "password", "sms_code"],
        "properties": {
          "phone": { "type": "string", "pattern": "^1[3-9]\\d{9}$", "description": "手机号" },
          "password": { "type": "string", "minLength": 8, "description": "密码（≥8 位）" },
          "sms_code": { "type": "string", "pattern": "^\\d{6}$", "description": "6 位短信验证码" },
          "nickname": { "type": "string", "maxLength": 32 }
        }
      },
      "LoginReq": {
        "type": "object",
        "required": ["account", "password"],
        "properties": {
          "account": { "type": "string", "description": "手机号或邮箱" },
          "password": { "type": "string" }
        }
      },
      "ChangePasswordReq": {
        "type": "object",
        "required": ["old_password", "new_password"],
        "properties": {
          "old_password": { "type": "string" },
          "new_password": { "type": "string", "minLength": 8 }
        }
      },
      "CreateOrderReq": {
        "type": "object",
        "required": ["items", "address_id"],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["sku_id", "quantity"],
              "properties": {
                "sku_id": { "type": "string" },
                "quantity": { "type": "integer", "minimum": 1 }
              }
            }
          },
          "address_id": { "type": "string" },
          "coupon_id": { "type": ["string", "null"] },
          "remark": { "type": "string", "maxLength": 200 }
        }
      },
      "TokenResp": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string", "description": "JWT，2 小时过期" },
          "refresh_token": { "type": "string", "description": "刷新 token，30 天过期" },
          "expires_in": { "type": "integer", "description": "access_token 秒数" }
        }
      },
      "UserResp": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "phone": { "type": "string" },
          "nickname": { "type": "string" },
          "avatar": { "type": "string", "format": "uri" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "OrderResp": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["PENDING", "PAID", "SHIPPED", "COMPLETED", "CANCELLED"] },
          "total_amount": { "type": "number" },
          "items": { "type": "array", "items": { "type": "object" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProductDetailResp": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "skus": { "type": "array", "items": { "type": "object" } },
          "stock": { "type": "integer" }
        }
      },
      "PageRespProduct": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ProductDetailResp" } }
        }
      },
      "PageRespOrder": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderResp" } }
        }
      }
    }
  }
}